+++ module-services/service-audio/api/AudioServiceAPI.hpp
@@ -40,9 +40,6 @@ namespace AudioServiceAPI
      *  Response will come as message AudioStartRoutingResponse
      */
     bool RoutingStart(sys::Service *serv);
-    bool RoutingMute(sys::Service *serv, bool enable);
-    bool RoutingSpeakerPhone(sys::Service *serv, bool enable);
-    bool RoutingHeadset(sys::Service *serv, bool enable);
     /**
      * @brief Stops playback operations by type. Asynchronous call.
      *
@@ -90,6 +87,22 @@ namespace AudioServiceAPI
      */
     bool Resume(sys::Service *serv, const audio::Token &token);
     /**
+     * @brief Sends audio event
+     * @param serv Requesting service.
+     * @param evt Event to be sent.
+     * @return True is request has been sent successfully, false otherwise
+     *   Response will come as message AudioSendEventResponse
+     */
+    bool SendEvent(sys::Service *serv, std::shared_ptr<audio::Event> evt);
+    /**
+     * @brief Sends audio event
+     * @param serv Requesting service.
+     * @param evt Event to be sent.
+     *  @return True is request has been sent successfully, false otherwise
+     *   Response will come as message AudioSendEventResponse
+     */
+    bool SendEvent(sys::Service *serv, audio::EventType evt);
+    /**
      * @brief Attempts to parse audio file for metatags.
      *
      * @param serv Requesting service.
@@ -127,7 +140,6 @@ namespace AudioServiceAPI
                               const T value,
                               const audio::PlaybackType &playbackType = audio::PlaybackType::None,
                               const audio::Profile::Type &profileType = audio::Profile::Type::Idle);
-
     /** @brief Key pressed handler.
      *
      * @param serv - requesting service.